#30135 Batch render upgrades to support 2016.1
In Flame 2016 extension 1, the batch render export profile interface
changed. This change upgrades the flame integration so that it supports
these changes in 2016.1.

The shot export xml now requires a new 'outputPathPattern' key which
defines where batch renders should go on disk.

This means that you can independently configure the location of the
initial rendered plates and the location of subsequent batch renders.

Closes #7.