Force aspect ratio to be divisible by 4
JIRA: SMOK-46515

DESCRIPTION

Exports from Wiretap Central are corrupted when the resolution of the
source sequence cannot be divided by 4.

The problem has also an impact on the Shotgun side as the Shotgun export
thumbnails are generated by the same tool used by Wiretap Central for
the exports (read_frame.bin).