upport movie asset type publish
JIRA: SMOK-48512
Support for movie asset in shotgun.

Following SMOK-48510, a new asset type (movie) can be exported.

We used to export it as video but since movies can also have audio,
it might be good to distinguish them.

userData passed by Flame's python hooks can now be a None object.
Must check and not assume that it is a dictionnary.